+++ jest.config.js
@@ -5,5 +5,7 @@ module.exports = {
   testMatch: ['<rootDir>/test/**/*.{spec,test}.{js,jsx,ts,tsx}'],
   // testMatch: ['<rootDir>/test/ctx/dispatch.test.js'],
   // testMatch: ['<rootDir>/test/api/'],
+  // testMatch: ['<rootDir>/test/api/dispatch.test.js'],
   // testMatch: ['<rootDir>/test/api-with-comp/'],
+  // testMatch: ['<rootDir>/test/api-with-comp/run.test.js'],
 };