+++ src/core/base/index.js
@@ -1,5 +1,4 @@
 export { default as boot } from './boot';
-export { default as catchCcError } from './catch-cc-error';
 export { default as computeCcUniqueKey } from './compute-cc-unique-key';
 export { default as runLater } from './run-later';
 export { default as uuid } from './uuid';