+++ src/core/computed/make-cu-ret-container.js
@@ -11,7 +11,6 @@ import { makeCuPackedValue, okeys, justWarning } from '../../support/util';
  */
 export default function (computed, originalCuContainer) {
   const moduleComputedValue = {};
-  
   okeys(computed).forEach(key => {
     // 用这个对象来存其他信息, 避免get无限递归，
     originalCuContainer[key] = makeCuPackedValue();