+++ test/api/dispatch.test.js
@@ -2,18 +2,19 @@ import { run, dispatch, getState, ccContext } from '../../src/index';
 import { makeStoreConfig, extractMessage } from '../util';
 
 const Foo = 'foo';
+const errorList = [];
 
 describe('test top api dispatch', () => {
   const models = makeStoreConfig(Foo);
   let fooState;
 
   beforeAll(() => {
-    run(models, { log: false });
+    run(models, { log: false, errorHandler: err => errorList.push(err) });
     fooState = getState(Foo);
   });
 
 
-  test('dispatch undeclared module', async () => {
+  test('dispatch undeclared module do nothing', async () => {
     const moduleCount = Object.keys(ccContext.moduleName2stateKeys).length;
     await dispatch(`xxx/changeName`, 'payload');
     const moduleCountAfterDispatch = Object.keys(ccContext.moduleName2stateKeys).length;
@@ -21,12 +22,16 @@ describe('test top api dispatch', () => {
   });
 
 
-  test('dispatch undeclared method', async () => {
-    try{
-      await dispatch(`${Foo}/notExist`, 'payload');
-    }catch(err){
-      expect(extractMessage(err)).toMatch(/(?=no reducer fn found)/);
-    }
+  test('dispatch undeclared method do nothing in unstrict mode', async () => {
+    await dispatch(`${Foo}/notExist`, 'payload');
+  });
+
+
+  test('dispatch undeclared method should throw error in strict mode', async () => {
+    ccContext.runtimeVar.isStrict = true; // hack
+    await dispatch(`${Foo}/notExist`, 'payload');
+    expect(errorList[0].message).toMatch(/(?=no reducer fn found)/);
+    ccContext.runtimeVar.isStrict = false;
   });